# TST: speed-up MLP tests by using default batch size