@jeremiedbb 2022-04-28
What I don't like with leaving batch_size to its default value is that all common tests use a very small dataset (n_samples=30 in most of them) which makes us only test the behavior when batch_size=n_samples. I think it's important to test the behavior when batch_size < n_samples.

I see that max_iter is set to 100 above for mlp. Can you try keeping batch_size to a small value, but lower the max_iter ?