@jeremiedbb 2022-04-29
> Common tests are not really meant to test convergence properties of estimators but more like the general behavior.

I agree and that's why I don't like testing against batch=n_samples. This is not the general case, it's a special case.